+++ workspace/marqua/src/artisan/brush.ts
@@ -2,20 +2,20 @@ import { getHighlighter } from 'shiki';
 import { escape, generate } from '../utils.js';
 
 export interface Dataset {
-	language?: string;
-	lineStart?: number;
-	title?: string;
+	lang?: string;
+	file?: string;
+	[data: string]: string | undefined;
 }
 
 export const highlighter = await getHighlighter({ theme: 'github-dark' });
 
 export function transform(source: string, dataset: Dataset) {
 	const { codeToThemedTokens } = highlighter;
-	const { language, title, lineStart = 1 } = dataset;
+	const { file, ...rest } = dataset;
 
 	let highlighted = '';
-	let line = lineStart;
-	for (const tokens of codeToThemedTokens(source, language)) {
+	let line = +(rest['line-start'] || 1);
+	for (const tokens of codeToThemedTokens(source, rest.lang)) {
 		let code = `<code data-line="${line++}">`;
 		for (const { content, color } of tokens) {
 			const style = color ? `style="color: ${color}"` : '';
@@ -24,13 +24,18 @@ export function transform(source: string, dataset: Dataset) {
 		highlighted += `${code}</code>\n`;
 	}
 
+	rest.language = dataset.lang || ''; // fallback for HTML attribute
+	const attrs = Object.entries(rest).map(
+		([k, v]) => `data-${k.toLowerCase().replace(/[^a-z\-]/g, '')}="${escape(v || '')}"`
+	);
+
 	// needs to be /^<pre/ to prevent added wrapper from markdown-it
 	return `<pre data-mrq="block" class="mrq">
-	<header 
+	<header
 		data-mrq="header"
-		data-language="${language || ''}"
-		class="mrq ${title ? '' : 'empty'}"
-	>${title ? `<span>${title}</span>` : ''}
+		${attrs.join('\n\t\t')}
+		class="mrq ${file ? '' : 'empty'}"
+	>${file ? `<span>${file}</span>` : ''}
 		<div data-mrq="toolbar" class="mrq">
 			${generate.icon('list', 'Toggle\nNumbering')}
 			${generate.icon('clipboard', 'Copy')}
@@ -39,8 +44,8 @@ export function transform(source: string, dataset: Dataset) {
 
 	<div
 		data-mrq="pre"
-		data-language="${language || ''}"
-		class="mrq language-${language || 'none'}"
+		${attrs.join('\n\t\t')}
+		class="mrq language-${rest.lang || 'none'}"
 	>${highlighted.trim()}</div>
 </pre>`;
 }