+++ workspace/marqua/src/core/index.ts
@@ -1,37 +1,10 @@
 import type { MarquaTable } from '../types.js';
-import { exists } from 'mauss/guards';
 import { generate } from '../utils.js';
-
-interface FrontMatterIndex {
-	[key: string]: string | string[] | FrontMatterIndex;
-}
-export function construct(raw: string): FrontMatterIndex {
-	const index: FrontMatterIndex = {};
-
-	for (const line of raw.split(/\r?\n/).filter(exists)) {
-		const match = /([:\w\d]+): (.+)/.exec(line.trim());
-		if (!match || (match && !match[2].trim())) continue;
-
-		const [key, data] = match.slice(1).map((g) => g.trim());
-		const array = data[0] === '[' && data[data.length - 1] === ']';
-		const pruned = array ? data.slice(1, -1).trim() : null;
-		const value = array ? (pruned ? pruned.split(',') : []) : data;
-
-		const [head, ...rest] = key.split(':');
-		index[head] = rest.length ? nest(rest, value, index[head]) : value;
-	}
-
-	return index;
-
-	function nest([parent, ...rest]: string[], data: string | string[], memo: any = {}): any {
-		return !parent ? data : { ...memo, [parent]: nest(rest, data, memo[parent]) };
-	}
-}
 
 export function parse(source: string) {
 	const match = /---\r?\n([\s\S]+?)\r?\n---/.exec(source);
-	const crude = source.slice(match ? (match.index || 0) + match[0].length + 1 : 0);
-	const memory = construct((match && match[1].trim()) || '');
+	const crude = source.slice(match ? match.index + match[0].length + 1 : 0);
+	const memory = construct((match && match[1].trim()) || '') as FrontMatter;
 
 	return {
 		content: inject(crude, memory),
@@ -84,6 +57,70 @@ export function parse(source: string) {
 			},
 		}),
 	};
+}
+
+type Primitives = null | boolean | string;
+type FrontMatter = { [key: string]: Primitives | Primitives[] | FrontMatter | FrontMatter[] };
+export function construct(raw: string, memo: Record<string, any> = {}): FrontMatter[string] {
+	const indent = indentation(raw);
+	if (!/[:\-\[\]|#]/gm.test(raw)) {
+		return indent > 1 ? dedent(raw) : coerce(raw.trim());
+	}
+	if (indent <= 1) raw = raw.trimStart();
+	if (/^(".*"|'.*')$/.test(raw.trim())) {
+		return raw.trim().slice(1, -1);
+	}
+
+	const PATTERN = /(^[^:\s]+):(?!\/)\n?([\s\S]*?(?=^\S)|[\s\S]*$)/gm;
+	let match: null | RegExpExecArray;
+	while ((match = PATTERN.exec(raw))) {
+		const [, key, value] = match;
+		const i = indentation(value);
+		const data = construct(i ? dedent(value) : value, memo[key]);
+		if (Array.isArray(data) || typeof data !== 'object') memo[key] = data;
+		else memo[key] = { ...memo[key], ...data };
+	}
+
+	if (Object.keys(memo).length) return memo;
+
+	const cleaned = raw.replace(/#.*$/gm, '').trim();
+	switch (cleaned[0]) {
+		case '-': {
+			const sequence = cleaned.split('-').filter((v) => v);
+			type Possibly = Primitives & FrontMatter; // what the ...?
+			return sequence.map((v) => construct(dedent(v)) as Possibly);
+		}
+		case '[': {
+			const pruned = cleaned.slice(1, -1);
+			return pruned.split(',').map(coerce);
+		}
+		case '|': {
+			return dedent(cleaned.slice(1).replace('\n', ''));
+		}
+		default: {
+			return coerce(cleaned.trim());
+		}
+	}
+}
+
+// ---- internal functions ----
+
+function coerce(u: string) {
+	const v = u.trim(); // argument can be passed as-is
+	const map = { true: true, false: false, null: null };
+	if (v in map) return map[v as keyof typeof map];
+	// if (!Number.isNaN(Number(v))) return Number(v);
+	return /^(".*"|'.*')$/.test(v) ? v.slice(1, -1) : v;
+}
+
+function dedent(input: string) {
+	const indent = indentation(input);
+	const lines = input.split(/\r?\n/);
+	return lines.map((l) => l.slice(indent)).join('\n');
+}
+
+function indentation(line: string) {
+	return (/^\s*/.exec(line) || [''])[0].length;
 }
 
 function inject(source: string, metadata: Record<string, any>) {